Fix on local fields access bug
 * Fix bug where non local field accesses were still detected as local
 accesses if they field was also present in the calling class
 * Remove redundant tests
 * Add tests for qualified field names
 * Remove redundant fieldUsageLocal, etc. from CKMethodResult
