+++ fixtures/fieldusage/FieldUsage.java
@@ -61,5 +61,15 @@ class FieldUsage {
 		return b + f2.a + f2.x;
 	}
 
+	public int m9() {
+		FieldUsage2 f2 = new FieldUsage2();
+		f2.a = 10;
+		return f2.a;
+	}
+
+	public void m10() {
+		this.a = 10;
+	}
+
 	private String xx;
 }
