+++ src/main/java/com/github/mauricioaniche/ck/CKVisitor.java
@@ -1278,6 +1278,11 @@ public class CKVisitor extends ASTVisitor {
 		}
 	}
 
+	public void endVisit(QualifiedName node) {
+		classes.peek().classLevelMetrics.stream().map(metric -> (CKASTVisitor) metric).forEach(ast -> ast.endVisit(node));
+		if(!classes.peek().methods.isEmpty())
+			classes.peek().methods.peek().methodLevelMetrics.stream().map(metric -> (CKASTVisitor) metric).forEach(ast -> ast.endVisit(node));
+	}
 	// TODO: add all other endVisit blocks
 
 }
