+++ src/main/java/com/github/mauricioaniche/ck/metric/NumberOfFields.java
@@ -1,13 +1,10 @@
 package com.github.mauricioaniche.ck.metric;
 
 import com.github.mauricioaniche.ck.CKClassResult;
-import com.github.mauricioaniche.ck.CKMethodResult;
 import org.eclipse.jdt.core.dom.FieldDeclaration;
 import org.eclipse.jdt.core.dom.Modifier;
 
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Map;
 import java.util.Set;
 
 import static com.github.mauricioaniche.ck.util.JDTUtils.getVariableName;
@@ -27,7 +24,6 @@ public class NumberOfFields implements CKASTVisitor, ClassLevelMetric {
 	@Override
 	public void visit(FieldDeclaration node) {
 		fields++;
-
 		fieldNames.add(getVariableName(node.fragments()));
 
 		boolean isPublic = Modifier.isPublic(node.getModifiers());
@@ -70,12 +66,5 @@ public class NumberOfFields implements CKASTVisitor, ClassLevelMetric {
 		result.setNumberOfDefaultFields(defaultFields);
 		result.setNumberOfFinalFields(finalFields);
 		result.setNumberOfSynchronizedFields(synchronizedFields);
-
-		//filter the fieldUsage to contain only local fields and set it in the method result
-		for(CKMethodResult method : result.getMethods()){
-			Map<String, Integer> fieldUsageLocal = new HashMap<>(method.getFieldUsage());
-			fieldUsageLocal.keySet().retainAll(fieldNames);
-			method.setFieldUsageLocal(fieldUsageLocal);
-		}
 	}
 }
